chore(queries): pg_stat_user_tables: remove optimization to compute size of the materialized views without pg_XXX_size
Now we skip tables that have an AccessExclusiveLock that prevents pg_XXX_size to be executed,
we can rollback the complex optimization and use the simple form with pg_XXX_size for materialized views.

Because we use JOIN, if a AccessExclusiveLock is on a table/MatView, the table will be missing from the query,
and the line will be missed from the export page.

We could use LEFT JOIN:
- the line for the table will be kept, with NULL values for size fields.
For now, we prefer having missing data rather than NaN number reported.