feat: raise alert on long running queries per user instead of single pid
This commit modifies the existing alert on long running queries to be
raised at the user level instead of the finer granularity of the pid
level.

The reason is that if a query holds a lock which blocks several queries,
we will raise an alert for each query on hold, therefore spamming our
alert channels and creating noise.

Reducing the granularity to the user will improve readability and also
understand quicker which apps and services are impacted.